requirements.txt
# Core dependencies for image processing and web framework
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0
python-dotenv>=0.19.0

# Image processing libraries
Pillow>=8.3.1
piexif>=1.1.3
opencv-python>=4.5.0
webcolors>=1.11.1

# ML/AI libraries
torch>=1.9.0
transformers>=4.11.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Additional utilities
pyyaml
lxml>=4.6.0

# Documentation (optional for testing)
hf-doc-builder>=0.5.0
mdx_truly_sane_lists